refactor: ForgotPassword schema
